+++ src/pages/addAccountStep/AddAccountStepPage.tsx
@@ -3,8 +3,7 @@ import { useState } from 'react';
 import Header from '@/shared/ui/Header';
 import { ArrowDown, ArrowLeft } from '@/shared/assets/svgs/icon';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
-import { Button, Input } from '@/shared/design-system/ui';
-import DescriptionField from '@/shared/ui/DescriptionField';
+import { Button, DescriptionField, Input } from '@/shared/design-system/ui';
 import { BoundaryError } from '@/shared/types/error.type';
 import usePutUpdateAccount from '@/features/expense-management/api/usePutUpdateAccount';
 import useDisclosure from './hooks/useDisclosure';
