+++ src/pages/createExpenseStep/CreateExpenseStepPage.tsx
@@ -2,10 +2,9 @@ import { useLayoutEffect, useRef, useState } from 'react';
 import { FormProvider } from 'react-hook-form';
 import { useLoaderData, useNavigate } from 'react-router';
 import { Close } from '@/shared/assets/svgs/icon';
-import { Button } from '@/shared/design-system/ui';
+import { Button, DescriptionField } from '@/shared/design-system/ui';
 import Header from '@/shared/ui/Header';
 import Text from '@/shared/ui/Text';
-import DescriptionField from '@/shared/ui/DescriptionField';
 import useAddExpenseFormArray from '@/features/expense-management/lib/useAddExpenseFormArray';
 import Modal from '@/shared/ui/Modal';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
