+++ src/pages/expenseDetail/ui/ExpenseTimeHeader/index.tsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
-import DescriptionField from '@/shared/ui/DescriptionField';
 import { Copy, Crown, DollarCircle } from '@/shared/assets/svgs/icon';
 import { useTheme } from 'styled-components';
 import Text from '@/shared/ui/Text';
 import { useLoaderData } from 'react-router';
 import Modal from '@/shared/ui/Modal';
 import copyClipboard from '@/shared/lib/copyClipboard';
-import { TextButton } from '@/shared/design-system/ui';
+import { DescriptionField, TextButton } from '@/shared/design-system/ui';
 import { showToast } from '@/shared/ui/Toast';
 import Flex from '@/shared/ui/Flex';
 import { useGetGroupHeader } from '@/features/settlement-details/api/useGetGroupHeader';
@@ -183,7 +182,6 @@ function ExpenseTimeHeader({
             </TextButton>
           </Flex>
         }
-        bgColor="semantic.background.normal.alternative"
       />
       <CurvedProgressBar percentage={percentage}>
         <S.ModdoButton onClick={handleModdoButtonClick}>
