+++ src/pages/qrStep/QrStepPage.tsx
@@ -6,8 +6,7 @@ import saveAs from 'file-saver';
 import { ArrowLeft, Reset } from '@/shared/assets/svgs/icon';
 import { showToast } from '@/shared/ui/Toast';
 import Header from '@/shared/ui/Header';
-import { Button } from '@/shared/design-system/ui';
-import DescriptionField from '@/shared/ui/DescriptionField';
+import { Button, DescriptionField } from '@/shared/design-system/ui';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
 import generateShareLink from '@/shared/lib/generateShareLink';
 import * as S from './QrStepPage.styles';
