+++ src/pages/shareStep/ShareStepPage.tsx
@@ -5,9 +5,8 @@ import LoginHamImg from '@/shared/assets/pngs/LoginHamImg.png';
 import { ArrowLeft } from '@/shared/assets/svgs/icon';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
 import { ROUTE } from '@/shared/config/route';
-import DescriptionField from '@/shared/ui/DescriptionField';
 import Header from '@/shared/ui/Header';
-import { Button } from '@/shared/design-system/ui';
+import { Button, DescriptionField } from '@/shared/design-system/ui';
 import ButtonGroup from '@/shared/ui/ButtonGroup';
 import Text from '@/shared/ui/Text';
 import initKakaoSDK from '@/shared/lib/initKakaoSDK';
